Crate marid

Marid

A process orchestration library

Modules

test_helpers
A testing module for Runner/Process traits.

Structs

Composer
The Composer will start each runner inside of its own thread when the run() function is called. The current behavior is an ordered setup/run, but in the future a parallel startup mode will be offered.

MaridProcess
Signifying the running state of a unit of work, a MaridProcess will spawn a new thread in order to not block the current thread.

Receiver
The receiving half of a channel.

Sender
The sending half of a channel.

Enums

ProcessError
Error type for a running Process.

Signal
The set of subscribable signals.

Traits

Process
A Process represents are running unit of work. It can be signaled and waited on.

Runner
A type implementing the Runner trait has the job of performing some arbitrary work while waiting for a signal indication shutdown. Upon receiving that defined shutdown Signal, the Runner must exit in a finite period of time.

Functions

launch
This function will start the specified runner as well as listen on the specified signals.

Type Definitions

MaridError
Error type for Marid Runners.

Implementations