Struct marid::MaridProcess

pub struct MaridProcess {
    // some fields omitted
}

Signifying the running state of a unit of work, a MaridProcess will spawn a new thread in order to not block the current thread.

Upon dropping, an instance of a MaridProcess will join on the running thread, potentially blocking.

Methods

impl MaridProcess

fn start(runner: Box<Runner + Send>, signaler: Sender<Signal>, recv: Receiver<Signal>) -> MaridProcess

Starts the specified runner with the given signal receiver.