Module marid::test_helpers

A testing module for Runner/Process traits.

When implementing different pieces of this crate, it was expedient to have some testing structs in place. This might also be useful for other developers who would want to ensure proper functionality.

Structs

TestError
Error type for testing

TestProcess
A test struct that implements the Process trait.

TestRunner
A test struct that implements the Runner trait.