Trait marid::Runner

pub trait Runner {
    fn run(self: Box<Self>, signals: Receiver<Signal>) -> Result<(), MaridError>;
    fn setup(&mut self) -> Result<(), MaridError>;
}

A type implementing the Runner trait has the job of performing some arbitrary work while waiting for a signal indication shutdown. Upon receiving that defined shutdown Signal, the Runner must exit in a finite period of time.

Required Methods

fn run(self: Box<Self>, signals: Receiver<Signal>) -> Result<(), MaridError>

The run function is called when a user wants to perform work.

fn setup(&mut self) -> Result<(), MaridError>

The setup function is called when a user wants to get ready to work.

This function should only complete once the type is ready to be run, and must complete in a finite period of time.